refactor(RacingCar) : Car의 멤버 변수 name을 클래스로 분리
