+++ src/test/java/racingcar/domain/CarTest.java
@@ -14,27 +14,6 @@ public class CarTest {
     private final NumberGenerator stoppableNumberGenerator = new StoppableNumberGenerator();
 
     @Test
-    void 이름_빈칸_실패() {
-        assertThatThrownBy(() -> new Car(("")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 이름 입력은 필수입니다.");
-    }
-
-    @Test
-    void 이름_공백_포함() {
-        assertThatThrownBy(() -> new Car(("1 2 3")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 이름에 공백이 존재합니다.");
-    }
-
-    @Test
-    void 이름_길이_초과() {
-        assertThatThrownBy(() -> new Car(("123456")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 이름은 5자 이하이어야 합니다.");
-    }
-
-    @Test
     void Car_생성_성공() {
         Car car = new Car("12345");
         assertThat(car.getName()).isEqualTo("12345");
