+++ src/test/java/racingcar/domain/CarsTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
 import org.junit.jupiter.api.Test;
 import racingcar.utils.numbergenerator.MovableNumberGenerator;
 import racingcar.utils.numbergenerator.NumberGenerator;
